requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gsheets==0.6.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
standard-imghdr==3.13.0
tornado==6.4.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.15
